Comparator added
