+++ src/main/java/ru/netology/tickets/TicketComparator.java
@@ -1,0 +1,11 @@
+package ru.netology.tickets;
+
+import java.util.Comparator;
+
+public class TicketComparator implements Comparator<Ticket> {
+
+    @Override
+    public int compare(Ticket o1, Ticket o2) {
+        return o1.getFlightTime() - o2.getFlightTime();
+    }
+}
